refactor(api): move updateUserEmailWithValidation controller method
to src/identity-access-management
