+++ api/lib/application/users/index.js
@@ -1,5 +1,6 @@
 import Joi from 'joi';
 
+import { userController as srcUserController } from '../../../src/identity-access-management/application/user/user.controller.js';
 import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
 import { SUPPORTED_LOCALES } from '../../../src/shared/domain/constants.js';
 import { EntityValidationError } from '../../../src/shared/domain/errors.js';
@@ -651,7 +652,7 @@ const register = async function (server) {
             assign: 'requestedUserIsAuthenticatedUser',
           },
         ],
-        handler: userController.updateUserEmailWithValidation,
+        handler: srcUserController.updateUserEmailWithValidation,
         validate: {
           params: Joi.object({
             id: identifiersType.userId,
